pyproject.toml
[project]
name = "prompt-regress"
version = "0.1.0"
description = "A cli tool for prompt regression testing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.2",
    "click>=8.2.1",
    "hf-xet>=1.1.3",
    "ollama>=0.5.1",
    "openai>=1.84.0",
    "pytest>=8.4.0",
    "pyyaml>=6.0.2",
    "sentence-transformers>=4.1.0",
    "tiktoken>=0.9.0",
]

[project.scripts]
prompt-regress = "prompt_regress.cli:main"
